// console/lib/api-client/generated/types.ts
export interface components {
  schemas: {
// ...
    LoginRequest: {
      /** Format: email */
      email: string;
      /** Format: password */
      password: string;
    };
// ...
  };
// ...
}